rustlings: variables: Complete variables
This section checks knowledge of how variables are declared with let
and are constant by default, typing, mut keyword for
mutability. Exercise 5 was the hardest because a variable was
redeclared as a different type and I wasn't supposed to change any of
the declarations, so I used shadowing. I put the new declaration in
its own block.